Move type validation to shared
