+++ showcode_forge/compilers/pytest_compiler.py
@@ -4,6 +4,7 @@ import importlib.util
 import inspect
 from functools import reduce
 from contextlib import contextmanager
+from ..shared import get_types
 
 @contextmanager
 def add_to_path(p):
@@ -34,37 +35,6 @@ def get_test_cases(module):
         raise Exception("Could not find test cases")
 
     return test_case_members[0].test_cases, test_case_members[0].parameters[0]
-
-
-def get_types(parameter):
-    if isinstance(parameter, str):
-        if(len(parameter) > 1):
-            return [1]
-        return [1, 2]
-    if isinstance(parameter, bool):
-        return [3]
-    if isinstance(parameter, int):
-        return [0, 5, 4]
-    if isinstance(parameter, float):
-        return [5, 4]
-
-    if not isinstance(parameter, (list, tuple)):
-        raise Exception("Unknown parameter type")
-
-    if len(parameter) == 0:
-        return [6, 7, 8, 9, 10, 11]
-    if all(isinstance(p, int) for p in parameter):
-        return [6, 11, 10]
-    if all(isinstance(p, (float, int)) for p in parameter):
-        return [11, 10]
-    if all(isinstance(p, bool) for p in parameter):
-        return [9]
-    if all(isinstance(p, str) and len(p) <= 1 for p in parameter):
-        return [8, 7]
-    if all(isinstance(p, str) for p in parameter):
-        return [7]
-
-    raise Exception("Unknown parameter array type")
 
 def build_parameter(name, inputs):
     possible_types = list(map(get_types, inputs))
